requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aiosqlite==0.15.0         # via -r requirements.in, tortoise-orm
async-lru==1.0.2          # via -r requirements.in
cffi==1.14.1              # via cryptg
cryptg==0.2.post1         # via -r requirements.in
iso8601==0.1.12           # via tortoise-orm
loguru==0.5.1             # via -r requirements.in
pyaes==1.6.1              # via telethon
pyasn1==0.4.8             # via rsa
pycparser==2.20           # via cffi, cryptg
pypika==0.37.16           # via tortoise-orm
python-dotenv==0.14.0     # via -r requirements.in
rsa==4.6                  # via telethon
telethon==1.16.0          # via -r requirements.in
tortoise-orm==0.16.14     # via -r requirements.in
typing-extensions==3.7.4.2  # via aiosqlite, tortoise-orm
